GoogleOAuthService: Stop using deprecated Google+ userinfo endpoint
Starting from March 7, Google is shutting down Google+ userinfo
endpoint: [1].

The Google+ Sign-in feature is fully deprecated and is being shut down
on March 7, 2019. This will be a progressive shutdown, with intermittent
failures starting as early as January 28, 2019. Developers should migrate
to the more comprehensive Google Sign-in authentication system.

Switch to using OAuth2 userinfo endpoint.

[1] https://developers.google.com/+/web/signin

Change-Id: I01f8b5816e39226da4fde2dcf66cc43670d94c08