build: Update build to work with bazel 0.25
Bazel 0.25 disabled transitive loads of dependencies
in .bzl files, so currently the build fails with bazel 0.25.

There are two ways to deal with this, either export the dependencies
or load the dependencies on site.

This fix opted for the first option as it seems to be the
preferred way by the current build layout.

Tested with bazel 0.24 and bazel 0.25 on OSX and Linux.

For more information about the change in bazel please
refer to https://github.com/bazelbuild/bazel/issues/5636.

Change-Id: Ibdecc0149ed5270e468a909b34c0c37169c7b0fd
Signed-off-by: Nikola Wuttke <heiligeslama@gmail.com>