Decode Keycloak JWTs as UTF-8
Prior to this, JWTs from Keycloak were being decoded using the platform default encoding.
At least in the Docker container, this is set to a non-UTF-8 encoding, which messes with names and potentially more.
RFC 7519 specifies in 7.1 that the message is "the UTF-8 representation of the JWT Claims Set", so it seems rather silly to not treat it as such.

Change-Id: I2dceaa48360024eef42b5467e2cec7e57c094267