Dex: Add support for signature verification using JWKS
This change adds support for verifying Dex tokens by querying a JWKS
URL endpoint. If no JWKS URL is provided in the configuration, the
existing behavior is maintained, where only the claims are returned as
the result. Additionally, this change includes an option to remove the
'/dex' prefix from the API (disabled by default). An option to link
existing Gerrit accounts is also added, following the implementation in
other provider classes.

Change-Id: I1b764d6a7af6125966a93e2939aba94fb0634cac