Dex: Add authentication support for Git-over-HTTP
This change introduces authentication support for Git-over-HTTP by
implementing the OAuthLoginProvider interface. The username used is the
same as the Gerrit username, and the password is a JSON object
containing the OAuth token. This token can be obtained through Git
credential helpers, as described in
https://git-scm.com/doc/credential-helpers. The password format is as
follows: {"id_token": "<jwt_token_contents>"}.

Change-Id: If1d21e40822c588ba0c4dba32ce44c26bfa58729